test for multiple freeways too
wrote test for computing aadt for multiple freeways.  output compared
against the result as computed in R.  Also include the R file for
doing it.

If it wasn't such a pain to get json into R, I'd probably just do this
all in R.

Actually, that isn't a bad idea, to parallelize this by calling lots
of R jobs.